refactor: 优化Parse代码结构
