+++ TYLDDB/Parser/V1.cs
@@ -9,7 +9,7 @@ namespace TYLDDB.Parser
     /// Data type parser.<br />
     /// 数据类型解析器。
     /// </summary>
-    public class DataParser
+    public class DataParser_V1
     {
         // 定义正则表达式来匹配键值对（类型::"key"="value"）
         private readonly static string pattern = @"(\w+)\s*::\s*""([^""]+)""\s*=\s*""([^""]+)"";";
